Fase_3/bomberos_django/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow>=10.2.0
reportlab==4.0.7
python-dateutil==2.8.2
